Add cron info in projects show
